create Map component
