+++ src/screens/Search.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react';
+import Map from '../components/Map';
 import SearchList from '../components/SearchList';
 import {Container, Header, Button, Card, Image, Icon, Grid, GridColumn, Segment, Menu, Table} from 'semantic-ui-react'
 import axios from "axios"
 
 const getResults = async () => {
     try {
-      return await axios.get('http://211.254.213.185:5000');
+      return await axios.get('http://211.254.213.185:5000/searchapi');
     } catch (error) {
       console.error(error);
     }
@@ -16,31 +17,6 @@ const showResults = async () => {
     
     console.log(results)
 };
-
-//--- 하단은 gis 관련 코드입니다. ---
-function markerDisplay(res, map){
-    res.data.DATA.some((element, idx) => {
-      if(element.xcode && element.ycode){
-        let point = latlngToUtmk(element.ycode, element.xcode)
-        let marker = new window.olleh.maps.overlay.Marker({
-          position: point,
-          caption: '공영주차장',
-          map: map
-        });
-        // marker.setIcon('https://www.gunpopark.or.kr/images/main/about_icon.png');
-        marker['attribute'] = element
-        marker.onEvent('click', function(e) {
-          window.olleh.infoWindowShow(this)
-        });
-      }
-    })
-  }
-
-function latlngToUtmk(x, y){ 
-    let p = window.olleh.maps.UTMK.valueOf(new window.olleh.maps.LatLng(x, y))
-    return p
-}
-//--- 상단은 gis 관련 코드입니다. ---
 
 const extra = (
     <div>
@@ -56,7 +32,7 @@ const CardExample = () => (
     }}
       image='https://cphoto.asiae.co.kr/listimglink/1/2017081115530600874_1.png'
       header='전시회 이름'
-      description='전시회에 대한 설명입니다. 전시회에 대한 설명입니다. '
+      description='전시회 장소'
       extra={extra}
     />
 )
@@ -75,21 +51,26 @@ const CardExample = () => (
       <Header textAlign="center" as='h2'>{message}</Header>
       <Grid columns={2}>
         <Grid.Column>
-            <div id='map_div' style={{
-                width:'100%',
-                height:'100%',
-                background:'red'
-            }}>
-                {extra}
-            </div>
+          <Map />
+            
         </Grid.Column>
 
         <Grid.Column>
             <div>
                 <Container>
+                  <Grid.Row>
+                    <div>
+                      <h1>추천 전시</h1>
+                    </div>
+                  </Grid.Row>
                     <Grid.Row>
                         <div>{CardExampleGroups()}</div>
                     </Grid.Row>
+                    <Grid.Row>
+                    <div>
+                      <h1>온라인 전시</h1>
+                    </div>
+                  </Grid.Row>
                     <Grid.Row>
                         <div>{CardExampleGroups()}</div>
                     </Grid.Row>
@@ -115,27 +96,8 @@ class Search extends React.Component {
         })
     }
 
-    async componentDidMount() {
-        var mapOpts = {
-            center: new window.olleh.maps.UTMK(960823.7, 1945435.52),
-            zoom: 8,
-            mapTypeId: 'ROADMAP'
-        };
-        var map = new window.olleh.maps.Map(document.getElementById("map_div",),
-            mapOpts
-        );
-
-        axios({
-            method:'get',
-            url : 'https://kt_map.gitlab.io/data/public_parking.json',
-            responseType: 'json',
-          }).then(res => {
-            markerDisplay(res, map);
-          }).catch(err => console.log(err));
-
- 
-        console.log(this.state.searchResultMessage)
-    }
+    // async componentDidMount() {
+    // }
 
     render() {
         console.log(this.state.searchResultMessage)
